+++ packages/pluggableWidgets/pie-doughnut-chart-native/src/PieDoughnutChart.editorConfig.ts
@@ -3,12 +3,16 @@ import { StructurePreviewProps } from "@mendix/piw-utils-internal";
 import { PieDoughnutChartPreviewProps } from "../typings/PieDoughnutChartProps";
 import PieWithLabels from "./assets/pie-w-labels.svg";
 import PieWithoutLabels from "./assets/pie-wo-labels.svg";
+import PieWithLabelsDark from "./assets/pie-w-labels-dark.svg";
+import PieWithoutLabelsDark from "./assets/pie-wo-labels-dark.svg";
 import DoughnutWithLabels from "./assets/doughnut-w-labels.svg";
 import DoughnutWithoutLabels from "./assets/doughnut-wo-labels.svg";
 
-export function getPreview(values: PieDoughnutChartPreviewProps): StructurePreviewProps {
+export function getPreview(values: PieDoughnutChartPreviewProps, isDarkMode: boolean): StructurePreviewProps {
+    const pieWithLabelsSVG = isDarkMode ? PieWithLabelsDark : PieWithLabels;
+    const pieWithoutLabelsSVG = isDarkMode ? PieWithoutLabelsDark : PieWithoutLabels;
     return values.presentation === "pie"
-        ? result(values.showLabels ? PieWithLabels : PieWithoutLabels)
+        ? result(values.showLabels ? pieWithLabelsSVG : pieWithoutLabelsSVG)
         : result(values.showLabels ? DoughnutWithLabels : DoughnutWithoutLabels);
 
     function result(svg: string): StructurePreviewProps {