+++ packages/pluggableWidgets/pie-doughnut-chart-native/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 - Dark theme icons for Tile and List view.
+- We added dark mode to structure mode preview.
 
 ## [1.0.0] - 2021-10-14